+++ lib/src/vector/float64x2_vector.g.dart
@@ -72,12 +72,18 @@ class Float64x2Vector with IterableMixin<double> implements Vector {
 
   Float64x2Vector.filled(
       this.length, num value, this._cache, this._simdHelper) {
-    _numOfBuckets = _getNumOfBuckets(length, _bucketSize);
-    final byteData = ByteData(_numOfBuckets * _bytesPerSimdElement);
-    _buffer = byteData.buffer;
+    if (length < 0) {
+      throw ArgumentError('Length cannot be negative');
+    }
 
-    for (var i = 0; i < length; i++) {
-      byteData.setFloat64(_bytesPerElement * i, value.toDouble(), Endian.host);
+    _numOfBuckets = _getNumOfBuckets(length, _bucketSize);
+    final list = Float64x2List(_numOfBuckets);
+    _buffer = list.buffer;
+
+    final simdValue = Float64x2.splat(value.toDouble());
+
+    for (var i = 0; i < _numOfBuckets; i++) {
+      list[i] = simdValue;
     }
   }